refactor: move main-only sync helpers
