+++ packages/insomnia/src/sync/__schemas__/type-schemas.ts
@@ -1,8 +1,15 @@
 import { createBuilder, type Schema } from '@develohpanda/fluent-builder';
 
 import { baseModelSchema } from '../../models/__schemas__/model-schemas';
-import type { BackendProject, Branch, MergeConflict, SnapshotStateEntry, StatusCandidate, Team } from '../types';
-import type { BackendProjectWithTeam } from '../vcs/normalize-backend-project-team';
+import type {
+  BackendProject,
+  BackendProjectWithTeam,
+  Branch,
+  MergeConflict,
+  SnapshotStateEntry,
+  StatusCandidate,
+  Team,
+} from '../types';
 
 export const projectSchema: Schema<BackendProject> = {
   id: () => 'id',
